move code to seomwhere better
